fix: Share common truncate and setup fixtures
